scripts/requirements.txt
docker
six
pyyaml
plumbum
doit
pandas
bitmath
tabulate
